Changed interface of UndoLogDiscard based on feedback.
Previously it took old pointer and size.  Now it just takes the new pointer.
In other words you call it with the address of the oldest byte you would
like to keep.

Also updated various comments, changed the pg_stat_undo_logs view a bit
and fixed the rules.out to reflect that.