Change the way that undolog.c accounts for page headers.
Previously, the offset component of UndoRecPtr was a counter of usable bytes
within the undo log.  Now it's straight physical offset into the undo log.
This means that it's now possible to have a (corrupt) UndoRecPtr that points
at header data, which is nonsensical, but it makes a lot of things simpler.
Now segment filenames, insert, discard, end offsets and UndoRecPtr values are
all based on the same scale: number of raw bytes from the start of the undo
log, regardless of whether those are header or data bytes.

Previously, UndoLogAllocate advanced the insert pointer by the exact number
of bytes requested in the 'size' argument.  That was unusable because it
required the caller to allocate header space this way too, and do a bunch of
math that required knowing where the insert point already was.  Now 'size'
means usable bytes; header bytes are automatically accounted for.

Based on a complaint from Dilip.  Let's try it this way and see if it makes
more sense.