pg_upgrade: Rename undo checkpoint file after pg_resetwal.
During pg_upgrade, pg_resetwal module is invoked which resets all WAL
related information in the Controlfile including the last checkpoint
LSN and then a server process is started. But, as the undo checkpoint
filename is based on last checkpoint LSN, the server startup process
fails as it is unable to find the undo checkpoint file corresponding
to the last checkpoint lsn as mentioned in the Controlfile.

This patch fixes that issue and also copies the undo data and checkpoint
files from old to new cluster during pg_upgrade.

TODO: Currently, this patch just adds the logic to copy undo data files
from default location i.e 'base/undo' which means if undo data files are
present in non-default location it won't work. In future, when the support
for storing undo data files in the non-default location is added, that
needs to be handled in pg_upgrade as well.

RM42357, Patch by me, reviewed by Thomas Munro.