Undo Action and Replay for Insert Operation
The undo insert action is to mark the corresponding item as dead if
relation has any index and unused otherwise.  The need to item as
dead is that we can't completely clear the item till the
corresponding index is marked as dead.  Currently undo actions are
replayed on Rollback or Rollback to savepoint.  Ideally the action
should be replayed on error as well, but we will deal with it in
a separate patch.

Amit Kapila, reviewed and tested by Rafia Sabih