GUC for storage_engine
Now, we have a GUC -- storage_engine to specify the storage_engine option,
which can be modified before server start. Currently, we have two options
for this GUC, viz. heap and zheap, with it's default value being heap.
Note that once this GUC is set to zheap, relations are created in
zheap format, irrespective of whether with (storage_engine ='zheap')
is specified in CREATE TABLE statement or not.

Reviewed by Amit Kapila.