Infrastructure to execute undo actions at page-level
Starting from the last undo record of a transaction read the undo
records in a backward direction till first undo record of a
transaction and accumulate the actions per page.  As soon as the
action is for a different page, we execute the accumulated actions
of previous page.

We are logging the complete page for undo actions, so we don't need
to record the data for individual operations.  We can optimize it by
recording the data for individual operations, but again if there are
multiple operations, then it might be better to log the complete
page.

Patch by Amit Kapila, reviewed by Rafia and Dilip