Rewind the undo pointer after applying undo actions
After applying all the undo actions for a page, we clear the
transaction slot on a page if the undo chain for block is complete,
otherwise rewind the undo pointer to the last record for that block
that precedes the last undo record for which action is replayed.

This will help us in knowing whether the undo action is already
replayed during recovery.

Patch by Amit Kapila, reviewed by Dilip Kumar