Support Epoch in Zheap pages
The idea is to make the change related to 64-bit transaction ids only
for the zheap pages whereas heap pages still operate with 32-bit
transaction ids. We will need wraparound and freeze vacuums for
metadata (system tables) stored in heap, but not for data stored in
zheap tables. The way to make 64-bit transaction ids in zheap is to
store epoch along with transaction id in each transaction slot (which
will make each transaction slot as 16 bytes (4 bytes transaction id,
4 bytes epoch, 8 bytes undo pointer)). Now, if we somehow ensure that
there is no undo for any transaction whose age is 2-billion years old
(wraparound limit), then we can easily make out the visibility using
current epoch and oldest_xid_having_undo. The way to achieve it is to
stop the system if it reaches such a situation. We piggybacked on the
existing wraparound machinery to raise different warning messages once
the system reaches that stage. If the epoch+xid in the page is lesser
than oldestXidWithEpochHavingUndo then it is all visible, otherwise,
the transaction will belong to current epoch and all the current rules
of current transaction system will work.

The reason for relying on 2-billion transaction age limit is that
current system (Transaction related functions like TransactionIdPrecedes)
relies on that and we don't want to change it.

We won't need any vacuum for freezing the transaction ids or wraparound
vacuums for zheap pages after this commit.

Amit Kapila, with some contribution by Dilip Kumar, reviewed and
verified by Kuntal Ghosh and Dilip Kumar.