Bug fix in rewind
If we are under a subtransaction then just reuse one
slot, because during the rollback of the subtransaction we will rewind the
undo insert location and the undo written for invalidating the slot will be
overwritten. So, it is better to invalidate only one slot which our
transaction is going to use.

Patch by Dilip Kumar reviewed by Amit Kapila