Move zheap_insert's buffer modifications inside critical section
Previously, it was not done because we thought we need to add the
tuple in page before forming the undo record as undo record requires
blockid and offset which we can only get after adding the tuple.
However, on closer inspection, actually, we only need blockid which we
can get from the buffer.

Patch by me, reviewed by Rafia Sabih