Fix crashes in page access after page pruning .
Issue is in many places while fetching or updating
tuple we failed to check if tuple has been
eleted/updated and pruned after we have released the
Buffer lock. This caused invalid access of pruned
tuples from the page. Now we now check if itemid is
deleted before accessing the tuple data in page.