Basic tablespace and persistence level support for undo logs.
Undo logs can now be created in different tablespaces, and can be created for
permanent and unlogged relations.  Undo logs for unlogged relations are
removed in the same circumstances that unlogged relations are.  The tablespace
used by each session is controlled by the GUC undo_tablespaces.  The columns
tablespace and persistence show this in the pg_stat_undo_logs view.
Tablespaces containing undo logs can be dropped and that automatically drops
the undo logs, but this is only allowed if no backend is attached and
currently in a transaction.

Warning: Some interlocking problems may remain, and recovery is still being
worked on.  Recovery and concurrent backup are still being worked on.  There
are known bugs in the checkpointing protocol that are being worked on.
Temporary relations appear to work superficially, but are not actually
temporary yet (they use shared buffers and their data is stored on disk
permanently).

Thomas Munro