Support Tid Scan for tables with zheap storage.
Use zheap equivalent of heap_fetch() to fetch the next tid.

To support WHERE CURRENT OF, have a zheap-equivalent of
heap_get_latest_tid() function. This function in turns uses
the zheap visibility function. For following the ctid chain
for non-in-place-updated tuples, have the MVCC visibility function
pass back the ctid of the new tuple.

Although zheap_get_latest_tid() accepts a snapshot, it may not work
with snapshots other than MVCC snapshot, because the corresponding
visibility functions for those snapshots are not modified to return
the new tuple ctid. This would be done in later commits, since it is
not necessary for Tid Scan.

Furthermore, the callers of heap_get_latest_tid()
(e.g. currtid_byreloid) should be modified to call
zheap_get_latest_tid() for zheap tables. This also would be done
in later commits.

Patch by Amit Khandekar, reviewed by Amit Kapila and Kuntal Ghosh.