Remove undo for INVALID_XACT_SLOT
Previously we needed this undo to identify the exact xid which
was there in the slot before reusing it.  And, that was stored
as uur_prevxid of the undo record.  Now, we are already including
uur_xid (transaction id which has inserted undo record) so we can
find the actual slot xid just by traversing the undo chain.

This will reduce the complexity of the code.  And, this will also
remove the limitation that one transaction is writing the undo in
other slots. So, by removing this limitation we can rewind the insert
location during rollback from the backend.

Patch by Dilip Kumar Review and Defect fixes by Amit Kapila