If full_page_writes is enabled, and the buffer image is not included in the WAL then we can rely on the tuple in the page to regenerate the undo tuple during recovery as the tuple state must be same as now, otherwise, we need to store it explicitly. But, in current code it is not stored externally even if the page image in included in the WAL.
Introduced new API XLogInsertExtended. Unlike XLogInsert, this function will
not retry for WAL insert if the page image inclusion decision got changed
instead it will return immediately. Also, it will not calculate the latest
value of the RedoRecPtr like XLogInsert does, instead it will take as input
from caller so that if the caller has decided to not to include the tuple
info (because page image is not present in the WAL) it can start over again
(if including page image decision got changed during WAL insertion). And
for the zheap wherever we need to include tuple info for generating the undo
record, we will call this new function instead of XLogInsert.

Patch by Dilip Kumar. Review and modified by Ashutosh Sharma and Amit Kapila